backend/pyproject.toml
[project]
requires-python = ">=3.10,<3.13"
name = "smart-dev-assistant-backend"
version = "0.1.0"
description = "FastAPI backend for SDA"
authors = [{ name = "Santosh Narayanan", email = "santosh33@gmail.com" }]

dependencies = [
    "fastapi",
    "pydantic",
    "python-dotenv",
    "psycopg2-binary",
    "sqlmodel",
    "qdrant-client",
    "sentence-transformers",
    "rank-bm25",
    "rapidfuzz",
    "uvicorn[standard]",
    "pillow",
    "python-multipart",
    "langchainhub",
    "langchain-experimental",
    "openai-whisper",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langchain-text-splitters",
    "httpx",
    "requests",
    "numpy",
    "scipy",
    "scikit-learn",
    "torch",
    "transformers",
    "tiktoken",
    "huggingface-hub",
    "anyio",
    "click",
    "jinja2",
    "packaging",
]

[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"